test(onboard): split custom Dockerfile coverage
